# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pecuaria API",
        "description": "Backend do sistema de gestao pecuaria (multi-fazenda) \u00e2\u20ac\u201d painel + motor de gatilhos.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/indicadores": {
            "get": {
                "tags": [
                    "indicadores"
                ],
                "summary": "Catalogo Indicadores",
                "operationId": "catalogo_indicadores_indicadores_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/IndicadorOut"
                                    },
                                    "type": "array",
                                    "title": "Response Catalogo Indicadores Indicadores Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/fazendas/{fazenda_id}/indicadores/valores": {
            "post": {
                "tags": [
                    "indicadores"
                ],
                "summary": "Lancar Valor",
                "description": "Lanca (ou atualiza) o valor de um indicador na fazenda e reavalia os gatilhos.",
                "operationId": "lancar_valor_fazendas__fazenda_id__indicadores_valores_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "fazenda_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Fazenda Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValorIn"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Lancar Valor Fazendas  Fazenda Id  Indicadores Valores Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/fazendas/{fazenda_id}/indicadores/{codigo}/serie": {
            "get": {
                "tags": [
                    "indicadores"
                ],
                "summary": "Serie Indicador",
                "description": "Serie temporal de um indicador (para graficos de tendencia).",
                "operationId": "serie_indicador_fazendas__fazenda_id__indicadores__codigo__serie_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "fazenda_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Fazenda Id"
                        }
                    },
                    {
                        "name": "codigo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Codigo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    },
                                    "title": "Response Serie Indicador Fazendas  Fazenda Id  Indicadores  Codigo  Serie Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FormatoIndicador": {
                "type": "string",
                "enum": [
                    "numero",
                    "percentual",
                    "moeda",
                    "dias"
                ],
                "title": "FormatoIndicador"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "IndicadorOut": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "codigo": {
                        "type": "string",
                        "title": "Codigo"
                    },
                    "nome": {
                        "type": "string",
                        "title": "Nome"
                    },
                    "categoria": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Categoria"
                    },
                    "unidade": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Unidade"
                    },
                    "formato": {
                        "$ref": "#/components/schemas/FormatoIndicador"
                    },
                    "casas_decimais": {
                        "type": "integer",
                        "title": "Casas Decimais"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "codigo",
                    "nome",
                    "categoria",
                    "unidade",
                    "formato",
                    "casas_decimais"
                ],
                "title": "IndicadorOut"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "ValorIn": {
                "properties": {
                    "indicador_codigo": {
                        "type": "string",
                        "title": "Indicador Codigo"
                    },
                    "valor": {
                        "type": "number",
                        "title": "Valor"
                    },
                    "data_ref": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Data Ref"
                    },
                    "observacao": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Observacao"
                    }
                },
                "type": "object",
                "required": [
                    "indicador_codigo",
                    "valor"
                ],
                "title": "ValorIn"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/auth/login"
                    }
                }
            }
        }
    }
}